+++ app/src/domain/entities/user.entity.ts
