+++ app/src/domain/exceptions/conflict.exception.ts
