+++ app/src/domain/exceptions/http.exception.ts
