+++ app/src/domain/exceptions/unauthorized.exception.ts
